requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.4.0
python-dotenv==1.0.0
pydantic==2.5.2
astroid